+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 **Changes in version 2.4**
 
-- The list of InputFields was updated for the following algorithms in the framework: NHIA, RUCA, URIC, Urban/Rural Continuum, Census Tract Poverty.  County at DX Reported (#90) was replaced with County at DX for Analysis (#89). 
+- Updated the Census Tract Poverty Indicator algorithm to handle 2015+ cases using the latest 5 year data from the ACS (2013-2017).
+- Added an algorithm to look up Purchased/Referred Care Delivery Areas (PRCDA), Urban Indian Health Organizations (UIHO), and UIHO Facility Numbers for a given state/county.
 
 **Changes in version 2.3**